refactored lib/bootstrap.js: most tests pass, still work to be done
