+++ lib/formSchemas/Blood_Labs_V2.js
@@ -1,0 +1,60 @@
+Meteor.startup(function() {
+  CRFprototypes.Blood_Labs_V2 = {
+    "Patient_ID": Patient_ID_Type,
+    "Sample_ID": Sample_ID_Type,
+    Day: { optional: true, type: 'Number' },
+    Phase: { optional: true, type: 'String' },
+    Segment: { optional: true, type: 'String' },
+    'TESTOSTERONE__TOTAL': { optional: true, type: 'Number' },
+    'Visit_Date': { optional: true, type: 'Date', autoform: autoformDate },
+    'Alkaline_Phosphatase': { optional: true, type: 'Number' },
+    'LDH_': { optional: true, type: 'Number' },
+    ANC: { optional: true, type: 'Number', decimal: true },
+    HEMATOCRIT: { optional: true, type: 'Number', decimal: true },
+    Hemoglobin: { optional: true, type: 'Number', decimal: true },
+    Platelets: { optional: true, type: 'Number' },
+    RBC: { optional: true, type: 'Number', decimal: true },
+    WBC: { optional: true, type: 'Number', decimal: true },
+    'WBC_Basophils': { optional: true, type: 'Number', decimal: true },
+    'WBC_Eosinophils': { optional: true, type: 'Number', decimal: true },
+    'WBC_Lymphocytes': { optional: true, type: 'Number', decimal: true },
+    'WBC_Monocytes': { optional: true, type: 'Number', decimal: true },
+    'WBC_Neutrophils': { optional: true, type: 'Number', decimal: true },
+    'Partial_Thromboplastin_Time____PTT_': { optional: true, type: 'Number', decimal: true },
+    'Prothrombin_Time__PT_': { optional: true, type: 'Number', decimal: true },
+    'PSA__complexed__direct_measurement_': { optional: true, type: 'Number', decimal: true },
+    Arm: { optional: true, type: 'String' },
+    'International_Normalized_Ratio__INR_': { optional: true, type: 'Number', decimal: true },
+    'Lab_Time': { optional: true, type: 'Number', decimal: true }
+  };
+
+  CRFfieldOrder.Blood_Labs_V2 = [
+    'Patient_ID',
+    'Sample_ID',
+    'Segment',
+    'Visit_Date',
+    'PSA__complexed__direct_measurement_',
+    'TESTOSTERONE__TOTAL',
+    'Alkaline_Phosphatase',
+    'LDH_',
+    'ANC',
+    'HEMATOCRIT',
+    'Hemoglobin',
+    'Platelets',
+    'RBC',
+    'WBC',
+    'WBC_Basophils',
+    'WBC_Eosinophils',
+    'WBC_Lymphocytes',
+    'WBC_Monocytes',
+    'WBC_Neutrophils',
+    'Partial_Thromboplastin_Time____PTT_',
+    'Prothrombin_Time__PT_',
+    'Arm',
+    'International_Normalized_Ratio__INR_',
+    'Lab_Time',
+    'Day',
+    'Phase',
+  ];
+
+});
