+++ lib/formSchemas/Demographics.js
@@ -1,0 +1,50 @@
+Meteor.startup(function() {
+  CRFprototypes.Demographics = {
+    "Patient_ID": Patient_ID_Type,
+    Age: { optional: true, type: 'Number' },
+    Arms: { optional: true, type: 'Number' },
+    'Birth_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Consent_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Current_Status': { optional: true, type: 'String' },
+    Ethnicity: { optional: true, type: 'String' },
+    'Expired_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    Gender: { optional: true, type: 'Number', decimal: true },
+    'Last_Visit': { optional: true, type: 'String' },
+    'Last_Visit_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Off_Study_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Off_Treatment_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'On_Followup_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'On_Study_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'On_Treatment_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    Race: { optional: true, type: 'String' },
+    'Study_Site': { optional: true, type: 'String' },
+    'Eligibility_Status': { optional: true, type: 'String' },
+    'Eligibility_Status_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Eligibility_Version_Date': { optional: true, type: 'Date', autoform: autoformDate  }
+  };
+
+  CRFfieldOrder.Demographics = [
+    'Patient_ID',
+    'Age',
+    'Study_Site',
+    // 'Birth_Date',
+    'Current_Status',
+    'Ethnicity',
+    'Expired_Date',
+    // 'Gender',
+    'Last_Visit',
+    'Last_Visit_Date',
+    'Off_Study_Date',
+    'Off_Treatment_Date',
+    'On_Followup_Date',
+    'On_Study_Date',
+    'On_Treatment_Date',
+    'Race',
+    'Eligibility_Status',
+    'Eligibility_Status_Date',
+    'Eligibility_Version_Date',
+    'Consent_Date',
+    'Arms',
+  ];
+
+});
