+++ lib/formSchemas/GU_Disease_Assessment_V3.js
@@ -14,22 +14,22 @@ Meteor.startup(function() {
     'Arm': { optional: true, type: 'String' },
     'Compared_with_previous_scan': { optional: true, type: 'String' },
     'Comments': { optional: true, type: 'String' }
-  }
+  };
 
   CRFfieldOrder.GU_Disease_Assessment_V3 =  [
-   'Patient_ID',
-   'Sample_ID',
-   'Are_Lesions_Present?',
-     'Date_of_Procedure',
-     'Date_of_Procedure_Ext',
-     'Day',
-     'Phase',
-     'Procedure',
-     'Segment',
-     'Visit_Date',
-     'Compared_with_previous_scan',
-     'Comments',
-     'Arm',
-   ]
+    'Patient_ID',
+    'Sample_ID',
+    'Are_Lesions_Present?',
+    'Date_of_Procedure',
+    'Date_of_Procedure_Ext',
+    'Day',
+    'Phase',
+    'Procedure',
+    'Segment',
+    'Visit_Date',
+    'Compared_with_previous_scan',
+    'Comments',
+    'Arm',
+  ];
 
 });
