+++ lib/formSchemas/SU2C_Prior_TX_V3.js
@@ -1,0 +1,62 @@
+Meteor.startup(function() {
+  CRFprototypes.SU2C_Prior_TX_V3 = {
+    "Patient_ID": Patient_ID_Type,
+    "Sample_ID": Sample_ID_Type,
+    "BL_PSA" : { optional: true, type: 'Number' },
+    "Bone_Response" : { optional: true, type: 'String' },
+    "PSA_Response" : { optional: true, type: 'String' },
+    "PSA_nadir" : { optional: true, type: 'Number' },
+    "PSA_nadir_Date" : { optional: true, type: 'Date', autoform: autoformDate  },
+    "PSA_nadir_Date_Ext" : { optional: true, type: 'Number' },
+    "Progression_Date" : { optional: true, type: 'Date', autoform: autoformDate  },
+    "Progression_Date_Ext" : { optional: true, type: 'Number' },
+    "RECIST_Response" : { optional: true, type: 'String' },
+     Day: { optional: true, type: 'Number' },
+    'Drug_Name': { optional: true, type: 'String' },
+    Phase: { optional: true, type: 'String' },
+    Segment: { optional: true, type: 'String' },
+    'Start_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Start_Date_Ext': { optional: true, type: 'String' },
+    'Stop_Date_Ext': { optional: true, type: 'String' },
+    'Treatment_Category': { optional: true, type: 'String' },
+    'Visit_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Stop_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'If_Progressive_Disease__Specify_Type': { optional: true, type: 'String' },
+    'Reason_for_Stopping_Treatment': { optional: true, type: 'String' },
+    'Reason_for_Stopping_Treatment_Details': { optional: true, type: 'String' },
+    'Treatment_Details': { optional: true, type: 'String' },
+    'If_other__specify': { optional: true, type: 'String' },
+    Arm: { optional: true, type: 'String' }
+  };
+
+  CRFfieldOrder.SU2C_Prior_TX_V3 = [
+    'Patient_ID',
+    'Sample_ID',
+    'Drug_Name',
+    'Treatment_Details',
+    "BL_PSA" ,
+    "Bone_Response" ,
+    "PSA_Response",
+    "PSA_nadir" ,
+    "PSA_nadir_Date" ,
+    "PSA_nadir_Date_Ext" ,
+    "Progression_Date",
+    "Progression_Date_Ext" ,
+    "RECIST_Response",
+    'Start_Date',
+    'Start_Date_Ext',
+    'Stop_Date',
+    'Stop_Date_Ext',
+    'If_Progressive_Disease__Specify_Type',
+    'Reason_for_Stopping_Treatment',
+    'Reason_for_Stopping_Treatment_Details',
+    'If_other__specify',
+    'Treatment_Category',
+    'Visit_Date',
+    'Phase',
+    'Segment',
+    'Arm',
+    'Day',
+  ];
+
+});
