+++ lib/formSchemas/SU2C_Subsequent_TX_V2.js
@@ -1,0 +1,54 @@
+Meteor.startup(function() {
+  CRFprototypes.SU2C_Subsequent_TX_V2 = {
+    "Patient_ID": Patient_ID_Type,
+    "Sample_ID": Sample_ID_Type,
+    Arm: { optional: true, type: 'String' },
+    Responder: { optional: true, type: 'String' },
+    ResponderEnzalutamide: { optional: true, type: 'String' },
+    ResponderAbiraterone: { optional: true, type: 'String' },
+    ResponderOtherTherapy: { optional: true, type: 'String' },
+    'Best_Response': { optional: true, type: 'String' },
+    Day: { optional: true, type: 'Number' },
+    'Drug_Name': { optional: true, type: 'String' },
+    'If_Progressive_Disease__Specify_Type': { optional: true, type: 'String' },
+    Phase: { optional: true, type: 'String' },
+    'Reason_for_Stopping_Treatment': { optional: true, type: 'String' },
+    Segment: { optional: true, type: 'String' },
+    'Start_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Start_Date_Ext': { optional: true, type: 'String' },
+    'Stop_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Stop_Date_Ext': { optional: true, type: 'String' },
+    'Treatment_Type': { optional: true, type: 'String' },
+    'Visit_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Treatment_Details': { optional: true, type: 'String' },
+    'Reason_for_Stopping_Treatment_Details': { optional: true, type: 'String' },
+    'If_other__specify': { optional: true, type: 'String' }
+  };
+
+  CRFfieldOrder.SU2C_Subsequent_TX_V2 = [
+    'Patient_ID',
+    'Sample_ID',
+    'Drug_Name',
+    'Reason_for_Stopping_Treatment',
+    'Responder',
+    'ResponderEnzalutamide',
+    'ResponderAbiraterone',
+    'ResponderOtherTherapy',
+    'Best_Response',
+    'If_Progressive_Disease__Specify_Type',
+    'Treatment_Type',
+    'Treatment_Details',
+    'Start_Date',
+    'Start_Date_Ext',
+    'Stop_Date',
+    'Stop_Date_Ext',
+    'Visit_Date',
+    'Reason_for_Stopping_Treatment_Details',
+    'If_other__specify',
+    'Day',
+    'Phase',
+    'Segment',
+    'Arm',
+  ];
+
+});
