+++ lib/formSchemas/SU2C_Subsequent_Treatment_V1.js
@@ -1,0 +1,74 @@
+Meteor.startup(function() {
+  CRFprototypes.SU2C_Subsequent_Treatment_V1 = {
+    "Patient_ID": Patient_ID_Type,
+    "Sample_ID": Sample_ID_Type,
+    Arm: { optional: true, type: 'String' },
+    Responder: { optional: true, type: 'String' },
+    ResponderEnzalutamide: { optional: true, type: 'String' },
+    ResponderAbiraterone: { optional: true, type: 'String' },
+    ResponderOtherTherapy: { optional: true, type: 'String' },
+    'BL_PSA':{ optional: true, type: 'Number' },
+    'PSA_nadir':{ optional: true, type: 'Number' },
+    'PSA_nadir_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'PSA_nadir_Date_Ext': { optional: true, type: 'String' },
+    'PSA_Response': { optional: true, type: 'String' },
+    'RECIST_Response': { optional: true, type: 'String' },
+    'Bone_Response': { optional: true, type: 'String' },
+    'Progression_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Progression_Date_Ext': { optional: true, type: 'String' },
+    'Resolved_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Best_Response': { optional: true, type: 'String' },
+    Day: { optional: true, type: 'Number' },
+    'Drug_Name': { optional: true, type: 'String' },
+    'If_Progressive_Disease__Specify_Type': { optional: true, type: 'String' },
+    Phase: { optional: true, type: 'String' },
+    'Reason_for_Stopping_Treatment': { optional: true, type: 'String' },
+    Segment: { optional: true, type: 'String' },
+    'Start_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Start_Date_Ext': { optional: true, type: 'String' },
+    'Stop_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Stop_Date_Ext': { optional: true, type: 'String' },
+    'Treatment_Category': { optional: true, type: 'String' },
+    'Visit_Date': { optional: true, type: 'Date', autoform: autoformDate  },
+    'Treatment_Details': { optional: true, type: 'String' },
+    'Reason_for_Stopping_Treatment_Details': { optional: true, type: 'String' },
+    'If_other__specify': { optional: true, type: 'String' }
+  };
+
+  CRFfieldOrder.SU2C_Subsequent_Treatment_V1 = [
+    'Patient_ID',
+    'Sample_ID',
+    'BL_PSA',
+    'PSA_nadir',
+    'PSA_nadir_Date',
+    'PSA_nadir_Date_Ext',
+    'PSA_Response',
+    'RECIST_Response',
+    'Bone_Response',
+    'Progression_Date',
+    'Progression_Date_Ext',
+    'Resolved_Date',
+    'Drug_Name',
+    'Reason_for_Stopping_Treatment',
+    'Responder',
+    'ResponderEnzalutamide',
+    'ResponderAbiraterone',
+    'ResponderOtherTherapy',
+    'Best_Response',
+    'If_Progressive_Disease__Specify_Type',
+    'Treatment_Category',
+    'Treatment_Details',
+    'Start_Date',
+    'Start_Date_Ext',
+    'Stop_Date',
+    'Stop_Date_Ext',
+    'Visit_Date',
+    'Reason_for_Stopping_Treatment_Details',
+    'If_other__specify',
+    'Day',
+    'Phase',
+    'Segment',
+    'Arm',
+  ];
+
+});
