+++ lib/formSchemas/newFormRequests/Nucleic_acid_isolation.txt
@@ -1,0 +1,32 @@
+Nucleic_acid_isolation form:
+    Source_material: ["Blood", "Frozen Tissue", "Fixed Tissue (FFPE)"]
+    Date of isolation
+    Blood case:
+        DNA:
+
+        Study_IDs: List of ( StudyID/Timepoint)
+        Sample_volume: 10-1000 microliters
+        Sample_concentration: 1-1000 nanograms per microliters
+        Storage location: Freezer_ID, Box_ID
+
+    Frozen or Fixed Tissue:
+        Isolation_type: ["RNA", "DNA"]
+        Study_IDs: List of ( StudyID/Core)
+        Samplename: StudyID / [RNA, RNA1, RNA2], StudyID / [DNA, DNA1, DNA2]
+        Elution volume: 10-100 microliters
+        Working aliquots: [1-3], Aliquots volume: 2-100 microliters
+        QC aliquot: YES/NO  Aliquots volume: 2-100 microliters
+        Storage location: freezerID, box ID
+
+Pathform:
+    TEST: positive, negative or failed
+
+    Ion torent: mutations
+
+RNA Seq completion forms ARE ALWAYS SAMPLES FROM LCM
+
+RNA Amplification
+RNA Library Prep
+
+Upload PDF for QC
+ Sample concentration
