--- run
@@ -1,5 +1,0 @@
-# export NODE_OPTIONS='--debug'
-# cd webapp
-export MONGO_URL=mongodb://localhost:27017/MedBook
-export ROOT_URL=http://localhost/CRF/
-meteor  --port 10002
