+++ run.sh
@@ -1,0 +1,6 @@
+# export NODE_OPTIONS='--debug'
+# cd webapp
+# export MONGO_URL=mongodb://localhost:27017/MedBook
+
+# mine
+ROOT_URL=http://localhost/CRF/ MONGO_URL=mongodb://localhost:27017/MedBook meteor --port 10002
