+++ tests/nightwatch/reports/FIREFOX_35.0_MAC_appLayout.xml
@@ -5,9 +5,9 @@
 
   <testsuite name="appLayout"
     errors="0" failures="0" hostname="" id="" package="appLayout" skipped="0"
-    tests="1" time="4.581" timestamp="Fri, 26 Jun 2015 19:34:57 GMT">
+    tests="1" time="5.354" timestamp="Tue, 30 Jun 2015 02:04:22 GMT">
   
-    <testcase name="Layout &amp; Static Pages" time="4.581" assertions="4">
+    <testcase name="Layout &amp; Static Pages" time="5.354" assertions="4">
 
     
 
@@ -18,7 +18,15 @@
     
     </testcase>
       <system-err>
-        Unable to locate element: &quot;table thead tr:nth-child(1) .Cores table tbody tr:nth-child(1) td:nth-child(1)&quot; using: css selector
+        Unable to locate element: &quot;table tbody tr:nth-child(1) .Study_Site&quot; using: css selector
+Unable to locate element: &quot;table tbody tr:nth-child(1) .Baseline_Sample_ID&quot; using: css selector
+Unable to locate element: &quot;table tbody tr:nth-child(1) .Baseline_Biopsy_Date&quot; using: css selector
+Unable to locate element: &quot;table tbody tr:nth-child(1) .Baseline_Biopsy_Site&quot; using: css selector
+Unable to locate element: &quot;table tbody tr:nth-child(1) .Study_Site&quot; using: css selector
+Unable to locate element: &quot;table tbody tr:nth-child(1) .Baseline_Sample_ID&quot; using: css selector
+Unable to locate element: &quot;table tbody tr:nth-child(1) .Baseline_Biopsy_Date&quot; using: css selector
+Unable to locate element: &quot;table tbody tr:nth-child(1) .Baseline_Biopsy_Site&quot; using: css selector
+Unable to locate element: &quot;table thead tr:nth-child(1) .Cores table tbody tr:nth-child(1) td:nth-child(1)&quot; using: css selector
 Unable to locate element: &quot;table thead tr:nth-child(1) .Cores table tbody tr:nth-child(2) td:nth-child(1)&quot; using: css selector
 Unable to locate element: &quot;table thead tr:nth-child(1) .Cores table tbody tr:nth-child(3) td:nth-child(1)&quot; using: css selector
 Unable to locate element: &quot;table thead tr:nth-child(1) .Cores table tbody tr:nth-child(4) td:nth-child(1)&quot; using: css selector
